+++ CHANGELOG.md
@@ -16,8 +16,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - 追加特性のカテゴリーセットにスケジュールを追加
 - 予約取引確定パラメータからobject(予約属性明示指定)を廃止
 - 興行オファー承認アクションのresultを最適化
-- 予約のITicketTypeを最適化
 - 施設コンテンツ、スケジュール、予約のproject属性を最適化
+- 予約のunderName,issuedByを最適化
 
 ### Deprecated